Allocators are allowed to return a larger memory chunk than was asked for. If the amount extra is large enough, then the hash table can use the extra space. The Global allocator will not hit this path, because it won't over-size enough to matter, but custom allocators may. An example of an allocator which allocates full system pages is included in the test suite (UNIX only because it uses `mmap`).

This implements #489. ~This relies on PR #524 to increase the minimum number of buckets for certain small types, which in turn constrains the domain of `maximum_buckets_in` so that the alignment can be ignored.~ Merged. 

~I haven't done any performance testing. Since this is on the slow path of making a new allocation, the feature should be doable without too much concern about overhead.~ There is a benchmark now you can run yourself, you can see the [numbers on my Apple M1 machine here](https://github.com/rust-lang/hashbrown/pull/523#issuecomment-3315091391). Since then, I added a fast-path shortcut which cuts this down when the system allocator is used (or when any allocator returns what was asked for).

I am definitely not an expert in swiss tables. Feedback is very welcome, even nitpicking.