@morrisonlevi 2024-05-07
So far I haven't been able to make it fail on real-world things. If you can make the `ctrl_align` quite a bit larger than the size, then the alignment becomes significant. It usually doesn't matter because:
1. Types are multiples of their align. This rules out things like `size: 128, ctrl_align: 4096`.
2. Bucket sizes are always powers of 2, and we never allocate less than 4 buckets.

Adjacent to this, I realize that on certain platforms, the minimum buckets being `4` is actually a smidge inefficient:

```
buckets: 4, TableLayout { size: 1, ctrl_align: 16 } leads to 36 bytes allocated
buckets: 8, TableLayout { size: 1, ctrl_align: 16 } leads to 40 bytes allocated
^-- using +4 more bytes more than doubles capacity (3 to 7)
```

Note that for the cost of `4` extra bytes, each of these becomes