@morrisonlevi 2024-05-08
I found an issue with very `TableLayout` sizes with large `ctrl_align`s. The optimization in #524 also corrects the domain of `maximum_buckets_in` because `buckets * table_layout.size` (or `x * y` above) will now be at least `ctrl_align`, so we can safely ignore it in the equation.