@Amanieu 2025-04-29
Could you write a quick benchmark to measure the added cost of allocating a new `HashMap`? Specifically just measuring `HashMap::with_capacity(X)` for different X values to see how much of an impact this has.