@morrisonlevi 2025-09-20
@Amanieu, I rebased my work, added a benchmark, and squashed my changes. This way anyone can check out the branch and compare to master by running:

```sh
# test this branch
cargo +nightly bench --bench with_capacity 

# revert the changes
git revert HEAD --no-edit

# test master
cargo +nightly bench --bench with_capacity 

# remove the revert commit to put us back to where we started
git reset --hard HEAD~1
```

This doesn't actually return the oversized check, since it's just using the system allocator. So the overhead being measured here is "the code performs the oversized check but the check doesn't result in a larger hashtable." Here are the formatted side-by-side results on my macOS M1:

Capacity | With feature (ns/iter) | Reverted (master) (ns/iter) | Overhead vs master
-- | -- | -- | --
0 | 3.50 ± 0.23 | 3.54 ± 0.05 | -1.1%
1 | 24.51 ± 0.60 | 21.18 ± 2.01 | +15.7%
3 | 24.62 ± 0.45 | 21.38 ± 1.67 | +15.2%
7 | 24.29 ± 0.71 | 21.83 ± 1.03 | +11.3%
8 | 51.60 ± 3.30 | 48.44 ± 3.45 | +6.5%
16 | 53.69 ± 2.72 | 51.28 ± 2.11 | +4.7%
32 | 20.60 ± 0.96 | 18.62 ± 0.68 | +10.6%
64 | 20.43 ± 0.44 | 19.18 ± 0.35 | +6.5%
128 | 21.88 ± 0.40 | 20.57 ± 0.37 | +6.4%
256 | 24.51 ± 1.71 | 25.11 ± 0.47 | -2.4%
512 | 31.11 ± 0.44 | 31.26 ± 0.47 | -0.5%
1,024 | 44.89 ± 4.70 | 44.21 ± 0.60 | +1.5%
4,096 | 156.25 ± 36.71 | 143.55 ± 40.47 | +8.9%
16,384 | 688.40 ± 30.35 | 685.02 ± 34.83 | +0.5%
65,536 | 14,748.30 ± 1,930.86 | 14,053.07 ± 1,588.73 | +5.0%

You can see results range from -2.4% to +15.7%. The results jive with what I would have suspected: the overhead is a bit higher on the smaller sizes because the overhead of the oversized check is roughly the same no matter the size, and for the larger sizes the other parts will dominate.

Note that for the allocators I personally have in mind, these sizes are fairly realistic. For example:
 1. Use a fixed-size of pre-allocated memory such as static memory (e.g. 64KiB).
 2. Round up to whole memory pages, which range in size from 4KiB to 16KiB depending on OS before including large pages.
 
I can check more sizes if you are interested, particular non-power-of-two sizes.